+++ src/commands/Minion/defaultfarming.ts
@@ -1,7 +1,7 @@
 import { CommandStore, KlasaMessage } from 'klasa';
 
-import { CompostTier } from '../../lib/farming/types';
 import { requiresMinion } from '../../lib/minions/decorators';
+import { CompostTier } from '../../lib/minions/farming/types';
 import { UserSettings } from '../../lib/settings/types/UserSettings';
 import { BotCommand } from '../../lib/structures/BotCommand';
 import { stringMatches } from '../../lib/util';
