+++ src/lib/minions/farming/defaultContracts.ts
