+++ src/lib/minions/farming/defaultPatches.ts
