+++ src/lib/types/minions.ts
@@ -1,7 +1,7 @@
 import { Peak } from '../../tasks/WildernessPeakInterval';
 import { Activity, Tasks } from '../constants';
-import { PatchTypes } from '../farming';
 import { MinigameIDsEnum } from '../minions/data/minigames';
+import { PatchTypes } from '../minions/farming';
 import { GroupMonsterActivityTaskOptions } from '../minions/types';
 import { BirdhouseData } from './../skilling/skills/hunter/defaultBirdHouseTrap';
